Changed the delete operation. Now it's performed passing only the id

+++ com.corbinelli.giamberini.examManagement/src/main/java/com/corbinelli/giamberini/examManagement/daos/CourseDAO.java
@@ -28,9 +28,12 @@ public class CourseDAO extends BaseDAO<Course> {
 	}
 
 	@Override
-	public void delete(Course course) {
+	public void delete(Long id) {
 		entityManager.getTransaction().begin();
-		entityManager.remove(course);
+		Course course = entityManager.find(Course.class, id);
+		if(course != null) {
+			entityManager.remove(course);
+		}
 		entityManager.getTransaction().commit();
 	}
