+++ com.corbinelli.giamberini.examManagement/src/main/java/com/corbinelli/giamberini/examManagement/daos/StudentDAO.java
@@ -27,9 +27,12 @@ public class StudentDAO extends BaseDAO<Student> {
 	}
 
 	@Override
-	public void delete(Student student) {
+	public void delete(Long id) {
 		entityManager.getTransaction().begin();
-		entityManager.remove(student);
+		Student student = entityManager.find(Student.class, id);
+		if(student != null) {
+			entityManager.remove(student);
+		}
 		entityManager.getTransaction().commit();
 	}
